modules building and working.  works only for Int32

+++ src/Modules/Math/ConvertMatrixToScalar.h
@@ -47,7 +47,7 @@ namespace Math {
     static const Dataflow::Networks::ModuleLookupInfo staticInfo_;
 
     INPUT_PORT(0, Input, Matrix);
-    OUTPUT_PORT(0, Output, Double);
+    OUTPUT_PORT(0, Output, Int32);
   };
 }}}
